requirements.txt
fastapi
uvicorn[standard]
requests
pyjwt
pymongo
tapipy
pyyaml
huggingface_hub
patra-toolkit